Allow users to override the startup timeout
On slower machines the one second startup timeout can sometimes not be
enough. This was fixed in the 1.5.3 Greenmail release (which the
project was recently upgraded to). This commit seeks to expose this new
functionality to the grails plugin configuration. It also corrects some
of the outdated information on the README.